+++ common/src/main/java/com/linecorp/centraldogma/internal/api/v1/ProjectDto.java
@@ -5,7 +5,7 @@
  * version 2.0 (the "License"); you may not use this file except in compliance
  * with the License. You may obtain a copy of the License at:
  *
- *   https://www.apache.org/licenses/LICENSE-2.0
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
@@ -14,10 +14,9 @@
  * under the License.
  */
 
-package com.linecorp.centraldogma.internal.httpapi.v1;
+package com.linecorp.centraldogma.internal.api.v1;
 
-import static com.linecorp.centraldogma.internal.httpapi.v1.HttpApiV1Constants.PROJECTS_PREFIX;
-import static com.linecorp.centraldogma.internal.httpapi.v1.HttpApiV1Constants.REPOS;
+import static com.linecorp.centraldogma.internal.api.v1.HttpApiV1Constants.PROJECTS_PREFIX;
 import static java.time.format.DateTimeFormatter.ISO_INSTANT;
 import static java.util.Objects.requireNonNull;
 
@@ -32,38 +31,29 @@ import com.google.common.base.MoreObjects;
 import com.google.common.base.MoreObjects.ToStringHelper;
 
 import com.linecorp.centraldogma.common.Author;
-import com.linecorp.centraldogma.common.Revision;
 
 @JsonInclude(Include.NON_NULL)
-public class RepositoryDto {
+public class ProjectDto {
 
     private final String name;
 
     private Author creator;
 
-    private Revision headRevision;
-
     private String url;
 
-    private String commitsUrl;
-
-    private String compareUrl;
-
-    private String contentsUrl;
+    private String reposUrl;
 
     private String createdAt;
 
-    public RepositoryDto(String name) {
+    public ProjectDto(String name) {
         this.name = requireNonNull(name, "name");
     }
 
-    public RepositoryDto(String projectName, String repoName, Author creator, Revision headRevision,
-                         long creationTimeMillis) {
-        this.name = requireNonNull(repoName, "repoName");
+    public ProjectDto(String name, Author creator, long creationTimeMillis) {
+        this.name = requireNonNull(name, "name");
         this.creator = requireNonNull(creator, "creator");
-        this.headRevision = requireNonNull(headRevision, "headRevision");
-        url = PROJECTS_PREFIX + '/' + requireNonNull(projectName, "projectName") + REPOS + '/' + repoName;
-        createdAt = ISO_INSTANT.format(Instant.ofEpochMilli(creationTimeMillis));
+        this.createdAt = ISO_INSTANT.format(Instant.ofEpochMilli(creationTimeMillis));
+        url = PROJECTS_PREFIX + '/' + name;
     }
 
     @JsonProperty("name")
@@ -78,33 +68,15 @@ public class RepositoryDto {
     }
 
     @Nullable
-    @JsonProperty("headRevision")
-    public Revision headRevision() {
-        return headRevision;
-    }
-
-    @Nullable
     @JsonProperty("url")
     public String url() {
         return url;
     }
 
     @Nullable
-    @JsonProperty("commitsUrl")
-    public String commitsUrl() {
-        return commitsUrl;
-    }
-
-    @Nullable
-    @JsonProperty("compareUrl")
-    public String compareUrl() {
-        return compareUrl;
-    }
-
-    @Nullable
-    @JsonProperty("contentsUrl")
-    public String contentsUrl() {
-        return contentsUrl;
+    @JsonProperty("reposUrl")
+    public String reposUrl() {
+        return reposUrl;
     }
 
     @Nullable
@@ -121,14 +93,10 @@ public class RepositoryDto {
             stringHelper.add("creator", creator());
         }
 
-        if (headRevision() != null) {
-            stringHelper.add("headRevision", headRevision());
-        }
-
         if (createdAt() != null) {
             stringHelper.add("createdAt", createdAt());
         }
-
         return stringHelper.toString();
     }
 }
+
