+++ common/src/main/java/com/linecorp/centraldogma/internal/api/v1/RepositoryDto.java
@@ -5,7 +5,7 @@
  * version 2.0 (the "License"); you may not use this file except in compliance
  * with the License. You may obtain a copy of the License at:
  *
- *   http://www.apache.org/licenses/LICENSE-2.0
+ *   https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
@@ -14,9 +14,10 @@
  * under the License.
  */
 
-package com.linecorp.centraldogma.internal.httpapi.v1;
+package com.linecorp.centraldogma.internal.api.v1;
 
-import static com.linecorp.centraldogma.internal.httpapi.v1.HttpApiV1Constants.PROJECTS_PREFIX;
+import static com.linecorp.centraldogma.internal.api.v1.HttpApiV1Constants.PROJECTS_PREFIX;
+import static com.linecorp.centraldogma.internal.api.v1.HttpApiV1Constants.REPOS;
 import static java.time.format.DateTimeFormatter.ISO_INSTANT;
 import static java.util.Objects.requireNonNull;
 
@@ -31,29 +32,38 @@ import com.google.common.base.MoreObjects;
 import com.google.common.base.MoreObjects.ToStringHelper;
 
 import com.linecorp.centraldogma.common.Author;
+import com.linecorp.centraldogma.common.Revision;
 
 @JsonInclude(Include.NON_NULL)
-public class ProjectDto {
+public class RepositoryDto {
 
     private final String name;
 
     private Author creator;
 
+    private Revision headRevision;
+
     private String url;
 
-    private String reposUrl;
+    private String commitsUrl;
+
+    private String compareUrl;
+
+    private String contentsUrl;
 
     private String createdAt;
 
-    public ProjectDto(String name) {
+    public RepositoryDto(String name) {
         this.name = requireNonNull(name, "name");
     }
 
-    public ProjectDto(String name, Author creator, long creationTimeMillis) {
-        this.name = requireNonNull(name, "name");
+    public RepositoryDto(String projectName, String repoName, Author creator, Revision headRevision,
+                         long creationTimeMillis) {
+        this.name = requireNonNull(repoName, "repoName");
         this.creator = requireNonNull(creator, "creator");
-        this.createdAt = ISO_INSTANT.format(Instant.ofEpochMilli(creationTimeMillis));
-        url = PROJECTS_PREFIX + '/' + name;
+        this.headRevision = requireNonNull(headRevision, "headRevision");
+        url = PROJECTS_PREFIX + '/' + requireNonNull(projectName, "projectName") + REPOS + '/' + repoName;
+        createdAt = ISO_INSTANT.format(Instant.ofEpochMilli(creationTimeMillis));
     }
 
     @JsonProperty("name")
@@ -68,15 +78,33 @@ public class ProjectDto {
     }
 
     @Nullable
+    @JsonProperty("headRevision")
+    public Revision headRevision() {
+        return headRevision;
+    }
+
+    @Nullable
     @JsonProperty("url")
     public String url() {
         return url;
     }
 
     @Nullable
-    @JsonProperty("reposUrl")
-    public String reposUrl() {
-        return reposUrl;
+    @JsonProperty("commitsUrl")
+    public String commitsUrl() {
+        return commitsUrl;
+    }
+
+    @Nullable
+    @JsonProperty("compareUrl")
+    public String compareUrl() {
+        return compareUrl;
+    }
+
+    @Nullable
+    @JsonProperty("contentsUrl")
+    public String contentsUrl() {
+        return contentsUrl;
     }
 
     @Nullable
@@ -93,10 +121,14 @@ public class ProjectDto {
             stringHelper.add("creator", creator());
         }
 
+        if (headRevision() != null) {
+            stringHelper.add("headRevision", headRevision());
+        }
+
         if (createdAt() != null) {
             stringHelper.add("createdAt", createdAt());
         }
+
         return stringHelper.toString();
     }
 }
-
