+++ server/src/main/java/com/linecorp/centraldogma/server/internal/admin/service/MetadataService.java
@@ -50,10 +50,10 @@ import com.linecorp.centraldogma.server.internal.admin.model.ProjectRole;
 import com.linecorp.centraldogma.server.internal.admin.model.RepoInfo;
 import com.linecorp.centraldogma.server.internal.admin.model.TokenInfo;
 import com.linecorp.centraldogma.server.internal.admin.model.UserAndTimestamp;
+import com.linecorp.centraldogma.server.internal.api.AbstractService;
 import com.linecorp.centraldogma.server.internal.command.Command;
 import com.linecorp.centraldogma.server.internal.command.CommandExecutor;
 import com.linecorp.centraldogma.server.internal.command.ProjectInitializer;
-import com.linecorp.centraldogma.server.internal.httpapi.AbstractService;
 import com.linecorp.centraldogma.server.internal.storage.project.Project;
 import com.linecorp.centraldogma.server.internal.storage.project.ProjectManager;
 import com.linecorp.centraldogma.server.internal.storage.project.ProjectNotFoundException;
