+++ server/src/main/java/com/linecorp/centraldogma/server/internal/admin/service/RepositoryService.java
@@ -59,9 +59,9 @@ import com.linecorp.centraldogma.server.internal.admin.dto.EntryWithRevisionDto;
 import com.linecorp.centraldogma.server.internal.admin.dto.RepositoryDto;
 import com.linecorp.centraldogma.server.internal.admin.dto.RevisionDto;
 import com.linecorp.centraldogma.server.internal.admin.exception.BadRequestException;
+import com.linecorp.centraldogma.server.internal.api.AbstractService;
 import com.linecorp.centraldogma.server.internal.command.Command;
 import com.linecorp.centraldogma.server.internal.command.CommandExecutor;
-import com.linecorp.centraldogma.server.internal.httpapi.AbstractService;
 import com.linecorp.centraldogma.server.internal.storage.project.ProjectManager;
 import com.linecorp.centraldogma.server.internal.storage.repository.FindOption;
 import com.linecorp.centraldogma.server.internal.storage.repository.Repository;
