+++ server/src/main/java/com/linecorp/centraldogma/server/internal/api/AbstractService.java
@@ -14,7 +14,7 @@
  * under the License.
  */
 
-package com.linecorp.centraldogma.server.internal.httpapi;
+package com.linecorp.centraldogma.server.internal.api;
 
 import static java.util.Objects.requireNonNull;
