+++ server/src/main/java/com/linecorp/centraldogma/server/internal/api/WatchService.java
@@ -14,7 +14,7 @@
  * under the License.
  */
 
-package com.linecorp.centraldogma.server.internal.httpapi;
+package com.linecorp.centraldogma.server.internal.api;
 
 import java.util.Collections;
 import java.util.Set;
