+++ server/src/test/java/com/linecorp/centraldogma/server/internal/api/CommitServiceV1Test.java
@@ -14,7 +14,7 @@
  * under the License.
  */
 
-package com.linecorp.centraldogma.server.internal.httpapi;
+package com.linecorp.centraldogma.server.internal.api;
 
 import static net.javacrumbs.jsonunit.fluent.JsonFluentAssert.assertThatJson;
