+++ server/src/test/java/com/linecorp/centraldogma/server/internal/api/WatchRequestConverterTest.java
@@ -14,7 +14,7 @@
  * under the License.
  */
 
-package com.linecorp.centraldogma.server.internal.httpapi;
+package com.linecorp.centraldogma.server.internal.api;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
@@ -30,7 +30,7 @@ import com.linecorp.armeria.common.HttpHeaderNames;
 import com.linecorp.armeria.common.HttpHeaders;
 import com.linecorp.armeria.server.ServiceRequestContext;
 import com.linecorp.centraldogma.common.Revision;
-import com.linecorp.centraldogma.server.internal.httpapi.WatchRequestConverter.WatchRequest;
+import com.linecorp.centraldogma.server.internal.api.WatchRequestConverter.WatchRequest;
 
 public class WatchRequestConverterTest {
